Adding ajax to project

+++ app/Http/Controllers/StudentController.php
@@ -7,31 +7,21 @@ use App\Models\Classes;
 use App\Models\Student;
 use App\Models\StudentParent;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
-use Illuminate\Validation\Rule;
-use Illuminate\View\View;
+use Illuminate\Support\Facades\Request;
 
 class StudentController extends Controller
 {
     public function index()
     {
+        if (request()->ajax()) {
+            return datatables()->of(Student::with('classes')->get())
+                ->addColumn('action', 'students.student-action')
+                ->rawColumns(['action'])
+                ->addIndexColumn()
+                ->make(true);
+        }
+        return view('students.index');
 
-        $students = Student::join('parents as p', 'parent', 'p.id')
-            ->join('classes as c', 'class', 'c.id')
-            ->orderBy('students.id')
-            ->select(
-                'students.*',
-                'c.name as c_name',
-                'p.first_name as p_fn',
-                'p.nni as p_nni',
-                'p.tel as p_tel',
-                'p.last_name as p_ln',
-                'p.sex as p_sex',
-                'p.date_of_birth as p_dob'
-            )
-            ->get();
-
-        return view('students.index')->with('students', $students);
     }
 
     public function create() {
@@ -84,6 +74,9 @@ class StudentController extends Controller
 
     public function view($id) {
         $student = $this->getStudentById($id);
+        if ($student == null) {
+            return redirect('notfound');
+        }
         return view('students.show', [
             'student' => $student
         ]);
@@ -91,6 +84,9 @@ class StudentController extends Controller
 
     public function edit($id) {
         $student = $this->getStudentById($id);
+        if ($student == null) {
+            return redirect('notfound');
+        }
         $classes = Classes::all();
         return view('students.create', [
             'student' => $student
@@ -133,12 +129,15 @@ class StudentController extends Controller
 
     }
 
-    public function destroy($id)
+    public function destroy()
     {
+        dd('hello');
         $student = $this->getStudentById($id);
         $parent = Student::select('parent', $student->p_id)
             ->where('parent', $student->p_id);
-
+        if ($student == null) {
+            return redirect('notfound');
+        }
         if ($parent->count() > 1) {
             $student->delete();
             return redirect()->back()->with('success', 'Suprimer avec succès');
@@ -152,27 +151,23 @@ class StudentController extends Controller
 
     private function getStudentById($id)
     {
-        $student = Student::join('parents as p', 'parent', 'p.id')
-            ->join('classes as c', 'class', 'c.id')
-            ->where('students.id',  $id)
-            ->orderBy('students.date_of_birth')
-            ->select(
-                'students.*',
-                'c.id as c_id',
-                'c.name as name',
-                'p.id as p_id',
-                'p.first_name as p_fn',
-                'p.nni as p_nni',
-                'p.tel as p_tel',
-                'p.last_name as p_ln',
-                'p.sex as p_sex',
-                'p.date_of_birth as p_dob'
-            )
-            ->first();
-
-        if ($student == null) {
-            return redirect('notfound');
-        }
+//        $student = Student::join('parents as p', 'parent', 'p.id')
+//            ->join('classes as c', 'class', 'c.id')
+//            ->where('students.id',  $id)
+//            ->select(
+//                'students.*',
+//                'c.id as c_id',
+//                'c.name as name',
+//                'p.id as p_id',
+//                'p.first_name as p_fn',
+//                'p.nni as p_nni',
+//                'p.tel as p_tel',
+//                'p.last_name as p_ln',
+//                'p.sex as p_sex',
+//                'p.date_of_birth as p_dob'
+//            )
+//            ->first();
+        $student = Student::with('classes', 'parents')->find($id);
         return $student;
     }
